Comments and Last page Responsive

+++ table-page/LastPage.js
@@ -1,9 +1,11 @@
 let abdAbsent = shimaaAbsent = tabarkAbsent = odayAbsent = amrooAbsent = abrarAbsent = ahmadAbsent = 1;
 let abdTask = shimaaTask = tabarkTask = odayTask = abrarTask = ahmadTask = amrooTask = 10;
-let abdSolvedTask = shimaaSolvedTask = tabarkSolvedTask = odaySolvedTask = abrarSolvedTask = ahmadSolvedTask = 9;
-let amrooSolvedTask = 5;
+let abdSolvedTask = shimaaSolvedTask = tabarkSolvedTask = odaySolvedTask = abrarSolvedTask = amrooSolvedTask = ahmadSolvedTask = 9;
 let flag = false;
 // 
+
+document.getElementById("image-of-user-M").style.display = 'none'
+document.getElementById("image-of-user-A").style.display = 'block'
 
 let arrayOfKeysAndValues = JSON.parse(localStorage.getItem('userInfo')) 
 let currentUser = JSON.parse(localStorage.getItem("current-user"));
@@ -82,7 +84,6 @@ document.getElementById("you-can-change-A").style.display = 'none'
 
 window.localStorage.setItem('student-data' , JSON.stringify(studentData))
 
-
 let callDataFromLocal = JSON.parse(localStorage.getItem('student-data'))
 
 document.getElementById("select").innerHTML = `<option value='first-select' selected>Select Developer</option>`
@@ -90,8 +91,6 @@ callDataFromLocal.forEach( function(element, index) {
     document.getElementById("select").innerHTML += `<option id='${index+1}' value='${index}' class='student'>ID ${index+1} : ${element.name}</option>`
 });
 
-
-console.log(document.getElementById('select').value);
 
 function setData() {
         document.getElementById("tableForNew").style.display = 'none'
@@ -126,8 +125,8 @@ function setData() {
         html += "<tr>" +  "<td>" + `  ID  `                     + "</td>" +    "<td class='center_table1'>" + callDataFromLocal[myValue].id + "</td>"  + "</tr>";
         html += "<tr>" +  "<td>" + `Student Name`               + "</td>" + "<td class='center_table1'>" + callDataFromLocal[myValue].name + "</td>"  + "</tr>";
         html += "<tr>" +  "<td>" + `Number of absent`           + "</td>" +   `<td id='AddAbsentHere' class='center_table1'> ${callDataFromLocal[myValue].absent}/180 <button id='clickToAddAbsent' onclick='increaseAbs(${callDataFromLocal[myValue].absent},${myValue})' > + </button>` + "</tr>";
-        html += "<tr>" +  "<td>" + `Number of Total Task`       + "</td>" +  `<td id='AddTaskHere' class='center_table1'>  ${callDataFromLocal[myValue].addTask} <button id='clickToAddTask' onclick='increaseTask(${callDataFromLocal[myValue].addTask},${myValue})' > + </button> <button id='clickToSubTask' onclick='decreaseTask(${callDataFromLocal[myValue].addTask},${myValue})'> - </button>  </td>` + "</tr>";
-        html += "<tr>" +  "<td>" + `Number of Task's Solved`    + "</td>" + `<td id='AddSolvedHere' class='center_table1'> ${callDataFromLocal[myValue].solvedTask} <button id='clickToAddSolved' onclick='increaseSolv(${callDataFromLocal[myValue].solvedTask},${myValue})' > + </button> <button id='clickToSubSolved' onclick='decreaseSolv(${callDataFromLocal[myValue].solvedTask},${myValue})'> - </button> </td>` + "</tr>";
+        html += "<tr>" +  "<td>" + `Number of Total Task`       + "</td>" +  `<td id='AddTaskHere' class='center_table1'>  ${callDataFromLocal[myValue].addTask} <button id='clickToAddTask' onclick='increaseTask(${callDataFromLocal[myValue].addTask},${myValue})' > + </button>` + "</tr>";
+        html += "<tr>" +  "<td>" + `Number of Task's Solved`    + "</td>" + `<td id='AddSolvedHere' class='center_table1'> ${callDataFromLocal[myValue].solvedTask} <button id='clickToAddSolved' onclick='increaseSolv(${callDataFromLocal[myValue].solvedTask},${myValue})' > + </button> ` + "</tr>";
         html += "<tr>" +  "<td>" + `Feedback from Coach Mona`   + "</td>" + `<td id='coach-mona-feedback' class='center_table1'> <p id='Mfeedbck'> ${callDataFromLocal[myValue].MFeedback} </p> </td>`  + "</tr>";
         html += "<tr>" +  "<td>" + `Feedback from Coach Alaa`   + "</td>" +  `<td id='coach-alaa-feedback' class='center_table1'> <p id='Afeedbck'> ${callDataFromLocal[myValue].Afeedback} </p> </td>` + "</tr>";
         html += "<tr>" +  "<td>" + `Delete Student`             + "</td>" +  `<td class='center_table1'> <button id='delete-student' class='trash' onclick='clickToRemoveStudent( ${myValue} )'> Remove <i class="fa-regular fa-trash-can"></i> </td> </button> ` + "</tr>"
@@ -153,28 +152,8 @@ function increaseTask(recieveData, indexValue){
         flag == false
     }
 }
-function decreaseTask(recieveData, indexValue){
-    recieveData = recieveData - 1;
-    studentData[indexValue].addTask = recieveData;
-    localStorage.setItem(`student-data` , JSON.stringify(studentData))
-    setData()
-    if (flag == true){
-        document.getElementById("clickToAddAbsent").style.setProperty("display", "none", "important")
-        flag == false
-    }
-}
 function increaseSolv(recieveData, indexValue){
     recieveData = recieveData + 1;
-    studentData[indexValue].solvedTask = recieveData;
-    localStorage.setItem(`student-data` , JSON.stringify(studentData))
-    setData()
-    if (flag == true){
-        document.getElementById("clickToAddAbsent").style.setProperty("display", "none", "important")
-        flag == false
-    }
-}
-function decreaseSolv(recieveData, indexValue){
-    recieveData = recieveData - 1;
     studentData[indexValue].solvedTask = recieveData;
     localStorage.setItem(`student-data` , JSON.stringify(studentData))
     setData()
@@ -204,53 +183,20 @@ function addNewStudent() {
     document.getElementById("you-can-change-M").style.display = 'none'
     document.getElementById("you-can-change-A").style.display = 'none'
     document.getElementById("tableForNew").style.display = 'block'
-    document.getElementById("tableContainer").style.setProperty("display", "none", "important") 
+    // document.getElementById("tableContainer").style.setProperty("display", "none", "important") 
     var addTable = "<table class='table2'  style='width:100%'>";  
     
-    addTable += "<tr>" + "<td>" + `ID`                         + "</td>" + `<td> ${callDataFromLocal.length + 1 || 1} </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Student Name`               + "</td>" + `<td id='newName'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Number of absent`           + "</td>" +  `<td id='numOfAbsent'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Number of Total Task`       + "</td>" + `<td id='numOfAbsent'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Number of Task's Solved`    + "</td>" + `<td id='numOfTasks'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Feedback from Coach Mona`   + "</td>" + `<td id='numOfSolved'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Feedback from Coach Alaa`   + "</td>" + `<td id='FeedBkFromM'>  </td>` + "</tr>";
-    addTable += "<tr>" + "<td>" + `Delete Student`             + "</td>" + `<td id='FeedBkFromA'>  </td>` + "</tr>";
-    
-
+    addTable += "<tr id='rowForId'>" + "<td>" + `ID for new Student`                         + "</td>" + `<td> ${callDataFromLocal.length + 1 || 1} </td>` + "</tr>";
     
     addTable += "</table>";
 
     addTable += `<table class='table3'>
                     <tr>
                         <td>
-                            <label>Add New Name :</label>
+                            <label id='labelMustBeHidden'>Add New Name :</label>
                         </td>
                         <td>
                             <input type='text' id='NN' value=''> *
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <label>Add Num Of Absent :</label>
-                        </td>
-                        <td>
-                            <input type='number' id='NA' value=''> *
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <label>Add Add Num Of Tasks :</label>
-                        </td>
-                        <td>
-                            <input type='number' id='NT' value=''> *
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <label>Add Add Num Of Solved :</label>
-                        </td>
-                        <td>
-                            <input type='number' id='NS' value=''> *
                         </td>
                     </tr>
                 </table>`
@@ -260,11 +206,11 @@ function addNewStudent() {
 }
 function addNewData(){
     newName = document.getElementById(`NN`).value;
-    newAbssent = document.getElementById(`NA`).value;
-    newTask = document.getElementById(`NT`).value;
-    newSolve = document.getElementById(`NS`).value;
+    newAbssent = 0
+    newTask = 0
+    newSolve = 0
 
-    if (newName != "" && newAbssent != "" && newTask != "" && newSolve != ""){
+    if (newName != ""){
         document.getElementById("select").innerHTML += `<option id='${callDataFromLocal.length + 1}' value='${callDataFromLocal.length + 1}' class='student'>${newName}</option>`
         let newStu = {
         id : callDataFromLocal.length + 1,
@@ -286,26 +232,30 @@ function addNewData(){
     } 
 }
 function clickToRemoveStudent(index){
-    let lengthOfSelectOption = document.getElementById("select").length-2
-    document.getElementById("tableContainer").style.setProperty("display", "none", "important")
-    document.getElementById("select").innerHTML = `<option value='first-select'>Select</option>`
-    for ( let i=0 ; i<(lengthOfSelectOption) ; i++ ){
-        if (i != index){
-        document.getElementById("select").innerHTML += `<option id='${i+1}' value='${i}' class='student'>ID ${i+1} : ${callDataFromLocal[i].name}</option>`
-    }else {
-        continue
-    }}
-    
-    let removeFromSotre = callDataFromLocal.forEach(function(element, indexOfObj){
-        if (indexOfObj == index){
-            callDataFromLocal.splice(index, 1)
-        }
-        localStorage.setItem(`student-data` , JSON.stringify(callDataFromLocal))
-    });
-    document.getElementById("you-can-change-A").style.display = 'none'
-    document.getElementById("you-can-change-M").style.display = 'none'
-    showAllStudent()
-    window.reload()
+    var result = confirm(`Ara you sure that you want to delete ${callDataFromLocal[index].name} from you list ?`)
+    if(result){
+        console.log(`${callDataFromLocal[index].name} is deleted`);
+        let lengthOfSelectOption = document.getElementById("select").length-2
+        document.getElementById("tableContainer").style.setProperty("display", "none", "important")
+        document.getElementById("select").innerHTML = `<option value='first-select'>Select</option>`
+        for ( let i=0 ; i<(lengthOfSelectOption) ; i++ ){
+            if (i != index){
+            document.getElementById("select").innerHTML += `<option id='${i+1}' value='${i}' class='student'>ID ${i+1} : ${callDataFromLocal[i].name}</option>`
+        }else {
+            continue
+        }}
+        
+        let removeFromSotre = callDataFromLocal.forEach(function(element, indexOfObj){
+            if (indexOfObj == index){
+                callDataFromLocal.splice(index, 1)
+            }
+            localStorage.setItem(`student-data` , JSON.stringify(callDataFromLocal))
+        });
+        document.getElementById("you-can-change-A").style.display = 'none'
+        document.getElementById("you-can-change-M").style.display = 'none'
+        showAllStudent()
+        window.reload()
+    }
 }
 function goToHome(){
     window.location.href = "../landing-page/landing-page.html"
@@ -343,6 +293,9 @@ function showAllStudent(){
     printAllStudent()
     document.getElementById("tableContainer").style.display = 'block'
     document.getElementById("show-all-student").style.visibility = 'hidden'
+    document.getElementById('select').value = 'first-select'
+    window.location.reload()
+    
 }
 
 function clearAll(){
@@ -383,17 +336,24 @@ function goToStudent(){
 if (currentUser.email == 'Alaa.amayreh2022@yahoo.com'){
     document.getElementById("image-of-user-A").src = "https://media-exp1.licdn.com/dms/image/C4E03AQFO1f75-n-0Xg/profile-displayphoto-shrink_800_800/0/1656781140615?e=1674086400&v=beta&t=oeJn8JyhW-MtKCxoq6qFg9VPgS5fh1ts2Wi6l_YWGxU"
     document.getElementById("user-name-home").innerHTML = currentUser.firstName + " " + currentUser.lastName
-    document.getElementById("email-of-user").innerHTML = currentUser.email
+    document.getElementById("image-of-user-M").style.display = 'none'
+    document.getElementById("image-of-user-A").style.display = 'block'
 
 } else if (currentUser.email == 'mona.salih@yahoo.com') {
     document.getElementById("image-of-user-M").src = "https://pps.whatsapp.net/v/t61.24694-24/174570944_302804968719494_5668236056779515453_n.jpg?ccb=11-4&oh=01_AdTRWcuJHJNWTd2a26nWslmv1XTH3e93fpRRorvrebYlCA&oe=63878EB1"
     document.getElementById("user-name-home").innerHTML = currentUser.firstName + " " + currentUser.lastName
-    document.getElementById("email-of-user").innerHTML = currentUser.email
+    document.getElementById("image-of-user-M").style.display = 'block'
+    document.getElementById("image-of-user-A").style.display = 'none'
 
 } else {
     document.getElementById("user-name-home").innerHTML = currentUser.firstName + " " + currentUser.lastName
-    document.getElementById("email-of-user").innerHTML = currentUser.email
 }
 function logOutNow(){
     window.location.href = "../index.html"
+}
+function confirmation(){
+    var result = confirm("Are you sure to delete?");
+    if(result){
+      console.log("Deleted")
+    }
 }
